pyproject.toml
[tool.poetry]
name = "interview-it"
version = "1.0.0"
description = "AI Based Interview Platform for IT Sector"
authors = ["Abdulmunim <66489217+abdulmunimjemal@users.noreply.github.com>"]
license = "Custom \"Read-Only\" License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.14"
langchain-openai = "^0.1.22"
langchain-community = "^0.2.12"
python-dotenv = "^1.0.1"
fastapi = {extras = ["all"], version = "^0.112.1"}
uvicorn = "^0.30.6"
requests = "^2.32.3"
deepgram-sdk = "^3.5.0"
aiohttp = "^3.10.5"
websockets = "^13.0"


[tool.poetry.group.dev.dependencies]
ipywidgets = "^8.1.3"
pydub = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
